Safari-611.1.10/LayoutTests/js/dom/modules/nomodule-dynamic-classic-inline.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test dynamic inlined classic script with nomodule.');
window.executed = false;

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
<script>
(function () {
    var element = document.createElement("script");
    element.noModule = true;
    element.textContent = `
        testFailed('error');
        window.executed = true;
    `;
    document.body.appendChild(element);
    setTimeout(function () {
        shouldBeFalse(`executed`);
        finishJSTest();
    }, 100);
} ());
</script>
</body>
</html>
