LayoutTests/js/dom/modules/nomodule-dynamic-classic-inline-expected.txt

Test dynamic inlined classic script with nomodule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS executed is false
PASS successfullyParsed is true

TEST COMPLETE
